requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.5
python-dotenv==1.0.1